backend/requirements.txt
# FastAPI y servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Pydantic
pydantic==2.4.2
pydantic-settings==2.0.3

# Base de datos
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# OpenAI
openai==1.3.5

# Redis
redis==5.0.1

# Logging
structlog==23.2.0

# IMPORTACIÓN DE EXCEL/CSV
numpy==1.26.0
pandas==2.1.0
openpyxl==3.1.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# Utilidades
email-validator==2.1.0
